+++ crates/config/src/config/mod.rs
@@ -3,6 +3,7 @@ mod errors;
 mod format;
 mod layer;
 mod loader;
+mod path;
 mod source;
 mod validator;
 
@@ -11,5 +12,6 @@ pub use errors::*;
 pub use format::*;
 pub use layer::*;
 pub use loader::*;
+pub use path::*;
 pub use source::*;
 pub use validator::*;